CRM-13872 - Improve idempotence
In this PR, we won't add a footnote if the URL already appears in the body.
When handling "git commit --amend" (or similar), this should prevent duplicate
footnotes.

This is not quite ideal -- it would be more effective to reparse the CommitMessage
and its footnotes then add any missing footnotes. The current implementation may
produce extraneous horizontal bars ("-----").

This does have the upside of making git-footnote a little less intrusive for
users who want to manually put in a URL.